Moved feed-list into its own package.
TODO: Add tests, and cleanup.
